formatting fix, file filter for encryption, changed encryption type
- % now displays correctly in progress bar
- added file filter for encrypt
- switched from block cipher to stream cipher

+++ src/dvsku_toolkit_gui/gui.cpp
@@ -14,7 +14,7 @@ void dvsku::toolkit::gui::handle_window_move() {
 	tagPOINT point;
 	GetCursorPos(&point);
 
-	glfwSetWindowPos(m_window, point.x - m_io.MouseClickedPos->x, point.y - m_io.MouseClickedPos->y);
+	glfwSetWindowPos(m_window, point.x - (int)m_io.MouseClickedPos->x, point.y - (int)m_io.MouseClickedPos->y);
 }
 
 void dvsku::toolkit::gui::build_gui() {
